+++ lib/modules/test/test_service.dart
@@ -1,9 +1,9 @@
 ﻿// Module: test
-// Updated at: 2026-07-03 17:06:36
-// Commit: perf(test): apply RepaintBoundary on animated test widgets
+// Updated at: 2026-07-03 17:14:35
+// Commit: test(test): implement widget integration tests for test
 
 class testComponentState {
-  final String commitRef = '2542';
-  final String lastMessage = 'perf(test): apply RepaintBoundary on animated test widgets';
-  final DateTime timestamp = DateTime.parse('2026-07-03T17:06:36+05:00');
+  final String commitRef = '2543';
+  final String lastMessage = 'test(test): implement widget integration tests for test';
+  final DateTime timestamp = DateTime.parse('2026-07-03T17:14:35+05:00');
 }